chore(moniker): move Moniker to keel-core
