+++ keel-clouddriver/keel-clouddriver.gradle.kts
@@ -5,6 +5,7 @@ plugins {
 
 dependencies {
   implementation(project(":keel-retrofit"))
+  implementation(project(":keel-core"))
   implementation("com.netflix.spinnaker.moniker:moniker")
   implementation("com.fasterxml.jackson.core:jackson-databind")
   implementation("com.fasterxml.jackson.core:jackson-annotations")
