+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/model/Moniker.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.spinnaker.keel.clouddriver.model
+package com.netflix.spinnaker.keel.model
 
 import com.fasterxml.jackson.annotation.JsonIgnore
