+++ keel-ec2-plugin/src/test/kotlin/com/netflix/spinnaker/keel/api/ec2/SecurityGroupTests.kt
@@ -3,10 +3,10 @@ package com.netflix.spinnaker.keel.api.ec2
 import com.netflix.spinnaker.keel.api.ec2.securityGroup.CidrRule
 import com.netflix.spinnaker.keel.api.ec2.securityGroup.PortRange
 import com.netflix.spinnaker.keel.api.ec2.securityGroup.ReferenceRule
-import com.netflix.spinnaker.keel.api.ec2.securityGroup.SecurityGroupRule.Protocol.TCP
 import com.netflix.spinnaker.keel.api.ec2.securityGroup.SecurityGroup
+import com.netflix.spinnaker.keel.api.ec2.securityGroup.SecurityGroupRule.Protocol.TCP
 import com.netflix.spinnaker.keel.api.ec2.securityGroup.SelfReferenceRule
-import com.netflix.spinnaker.keel.clouddriver.model.Moniker
+import com.netflix.spinnaker.keel.model.Moniker
 import de.danielbechler.diff.ObjectDifferBuilder
 import de.danielbechler.diff.node.DiffNode
 import de.danielbechler.diff.node.DiffNode.State.CHANGED
